chore: shared로 이동(프로필-마이페이지에서 사용)
